Fixed TemperatureTransmit payload & moved SensorData
- Fixed an issue where the payload wasn't empty before writing to it
- Fixed bitwise operations for payload
- Moved SensorData to a new namespace called DataModels
