+++ SEP4-IOT/SEP4-IOT.cproj
@@ -150,6 +150,12 @@
     <Compile Include="Config\LoraWAN_Config.h">
       <SubType>compile</SubType>
     </Compile>
+    <Compile Include="DataModels\SensorData.c">
+      <SubType>compile</SubType>
+    </Compile>
+    <Compile Include="DataModels\SensorData.h">
+      <SubType>compile</SubType>
+    </Compile>
     <Compile Include="drivers\display_7seg.h">
       <SubType>compile</SubType>
     </Compile>
@@ -291,12 +297,6 @@
     <Compile Include="FreeRTOS\src\variantHooks.c">
       <SubType>compile</SubType>
     </Compile>
-    <Compile Include="SensorData.c">
-      <SubType>compile</SubType>
-    </Compile>
-    <Compile Include="SensorData.h">
-      <SubType>compile</SubType>
-    </Compile>
     <Compile Include="Tasks\DataCollection.c">
       <SubType>compile</SubType>
     </Compile>
@@ -318,6 +318,8 @@
     <Folder Include="FreeRTOS\FreeRTOSTraceDriver\" />
     <Folder Include="FreeRTOS\src\" />
     <Folder Include="Config" />
+    <Folder Include="DataModels" />
+    <Folder Include="PlantUML" />
     <Folder Include="Tasks" />
   </ItemGroup>
   <ItemGroup>
